requirements.txt
streamlit
torch
openai-whisper
librosa
soundfile
matplotlib
numpy
